pyproject.toml
[tool.poetry]
name = "tiny-cloud-cli"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standart"], version = "^0.26.0"}
pyjwt = "^2.8.0"
python-multipart = "^0.0.6"
pymongo = {extras = ["srv"], version = "^4.6.1"}
bcrypt = "^4.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
